requirements.txt
# Python 3.10.5
Django==4.1
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
Pillow==9.1.1
tzlocal==4.2
babel==2.10.3
requests==2.28.0
selenium==4.2.0
django-environ==0.9
webdriver-manager==3.8.0
whitenoise==6.2.0